refactor: move classes into appropriate packages
