+++ src/main/java/me/qamulex/easyratelimiter/wrapper/RateLimiterMap.java
@@ -3,13 +3,15 @@
  *
  * Copyright (c) qamulex
  */
-package me.qamulex.easyratelimiter;
+package me.qamulex.easyratelimiter.wrapper;
 
 import java.util.Collection;
 import java.util.Map;
 import java.util.Set;
 
 import lombok.RequiredArgsConstructor;
+import me.qamulex.easyratelimiter.RateLimiter;
+import me.qamulex.easyratelimiter.RateLimiterBuilder;
 
 @RequiredArgsConstructor
 public class RateLimiterMap<K> implements Map<K, RateLimiter> {
